Add timeout to the user program run.
The user program is now timeouted when running longer than 5 seconds.
The value is set from environment variable, and is expected to be
configurable. The timeout command is used for this.

Also, some minor changes were made, like limiting the stderr and stdout
size to 10kb and adding overflow to the stdout/err boxes.