Change the web to be code executor only.
The web was multi-purpose. Now it only server for code running and
possibly some other gadgets in the future (like only diff, base64 and so
on).

The CSS of the page was also changed to reflect this change, now it
expects the editor to be there, and nothing else. The editor together
with the output is therefore 100% page width.