feat: Replace raw JS, HTML + Flask for Sveltekit and Tailwind.
The frontend part now uses the Svelte and Sveltekit. Additional proxy
was added between the frontend. The svelte migration unfortunately means
that some features were lost, like the automatic compilation. This is up
for future implementation. Since the app probably won't be run in k8s
for some time, the docker-compose was made "production" ready.