feat: Add metrics for each language, test and do not fail in statserver.
Add metrics for each language in evaluator. For each request and each
error in evaluator metric is collected.

Statserver now does not abort when unable to fetch metrics from one of
the endpoint. We should however plug some tests or alerts for this.

Add one small test for collecting metrics. Should probably be expanded
in the future.