feat: add Racket.
Add Racket to interpreted languages. Racket is, similarly to lua, built
from source. However, Racket build takes a very long time. As it is part
of evaluator build, the evaluator now builds **significantly** longer,
and this is something that we will have to fix in the future. Maybe
by building these images before hand and just pulling them from
repository?